+++ connectors/discord.ts
@@ -31,6 +31,7 @@ import {
   shouldShowToolOutput,
   extractImagePaths,
   removeImageMarkers,
+  stripThinkBlocks,
   sanitizeServerPaths,
 } from "../src"
 import { getConfig } from "../src/config"
@@ -333,7 +334,7 @@ class DiscordConnector extends BaseConnector<ChannelSession> {
       }
 
       // Clean response and send
-      const cleanResponse = sanitizeServerPaths(removeImageMarkers(responseBuffer))
+      const cleanResponse = sanitizeServerPaths(removeImageMarkers(stripThinkBlocks(responseBuffer)))
       if (cleanResponse) {
         session.outputChars += cleanResponse.length